Move pages from components directory to pages directory
