+++ expense_tracker_frontend/src/App.tsx
@@ -1,28 +1,28 @@
 import {BrowserRouter, Routes, Route, Navigate} from "react-router-dom";
-import {Accounts} from "./components/Accounts/Accounts";
-import {Tags} from "./components/Tags/Tags";
+import {Accounts} from "./pages/Accounts/Accounts";
+import {Tags} from "./pages/Tags/Tags";
 import "./components/common.scss";
-import {TransactionsList} from "./components/Transactions/Transactions";
-import {PresetsList} from "./components/Presets/Presets";
-import {Login} from "./components/Auth/Login";
-import {UserEdit} from "./components/Auth/UserEdit";
-import {TagCreate} from "./components/Tags/TagCreate";
-import {AccountCreate} from "./components/Accounts/AccountCreate";
-import {Tag} from "./components/Tags/Tag";
-import {TagDelete} from "./components/Tags/TagDelete";
-import {TagEdit} from "./components/Tags/TagEdit";
-import {Account} from "./components/Accounts/Account";
-import {AccountDelete} from "./components/Accounts/AccountDelete";
-import {AccountEdit} from "./components/Accounts/AccountEdit";
-import {PresetCreate} from "./components/Presets/PresetCreate";
-import {Preset} from "./components/Presets/Preset";
-import {PresetDelete} from "./components/Presets/PresetDelete";
-import {PresetEdit} from "./components/Presets/PresetEdit";
-import {TransactionCreate} from "./components/Transactions/TransactionCreate";
-import {Transaction} from "./components/Transactions/Transaction";
-import {TransactionDelete} from "./components/Transactions/TransactionDelete";
-import {TransactionEdit} from "./components/Transactions/TransactionEdit";
-import {AccountSync} from "./components/Accounts/AccountSync";
+import {TransactionsList} from "./pages/Transactions/Transactions";
+import {PresetsList} from "./pages/Presets/Presets";
+import {Login} from "./pages/Auth/Login";
+import {UserEdit} from "./pages/Auth/UserEdit";
+import {TagCreate} from "./pages/Tags/TagCreate";
+import {AccountCreate} from "./pages/Accounts/AccountCreate";
+import {Tag} from "./pages/Tags/Tag";
+import {TagDelete} from "./pages/Tags/TagDelete";
+import {TagEdit} from "./pages/Tags/TagEdit";
+import {Account} from "./pages/Accounts/Account";
+import {AccountDelete} from "./pages/Accounts/AccountDelete";
+import {AccountEdit} from "./pages/Accounts/AccountEdit";
+import {PresetCreate} from "./pages/Presets/PresetCreate";
+import {Preset} from "./pages/Presets/Preset";
+import {PresetDelete} from "./pages/Presets/PresetDelete";
+import {PresetEdit} from "./pages/Presets/PresetEdit";
+import {TransactionCreate} from "./pages/Transactions/TransactionCreate";
+import {Transaction} from "./pages/Transactions/Transaction";
+import {TransactionDelete} from "./pages/Transactions/TransactionDelete";
+import {TransactionEdit} from "./pages/Transactions/TransactionEdit";
+import {AccountSync} from "./pages/Accounts/AccountSync";
 
 function App() {
   return (
