+++ expense_tracker_frontend/src/components/Navbar.tsx
@@ -1,6 +1,6 @@
 import React, {useState} from "react";
 import {observer} from "mobx-react-lite";
-import {useToken} from "./Auth/AuthContext";
+import {useToken} from "../utils/AuthContext";
 import {AuthAxios} from "../utils/Network";
 import {Navbar, Container, Nav, NavDropdown} from "react-bootstrap";
