+++ expense_tracker_frontend/src/pages/Accounts/Account.tsx
@@ -1,15 +1,15 @@
 import {observer} from "mobx-react-lite";
-import {useToken} from "../Auth/AuthContext";
+import {useToken} from "../../utils/AuthContext";
 import React, {useEffect, useState} from "react";
 import {Link, useNavigate, useParams} from "react-router-dom";
-import {NavbarComponent} from "../Navbar";
-import {TableButton} from "../TableButton";
-import {StaticField} from "../StaticField";
-import {getSubtransactionBalances} from "../getSubtransactionBalances";
-import {centsToString, formatDate} from "../Tools";
+import {NavbarComponent} from "../../components/Navbar";
+import {TableButton} from "../../components/TableButton";
+import {StaticField} from "../../components/StaticField";
+import {getSubtransactionBalances} from "../../components/getSubtransactionBalances";
+import {centsToString, formatDate} from "../../components/Tools";
 import {AuthAxios} from "../../utils/Network";
 import {Col, Dropdown, Row, Container, Table} from "react-bootstrap";
-import {TimezoneTag} from "../TimezoneTag";
+import {TimezoneTag} from "../../components/TimezoneTag";
 import dayjs, {Dayjs} from "dayjs";
 
 interface AccountElement {
