+++ expense_tracker_frontend/src/pages/Accounts/AccountDelete.tsx
@@ -1,26 +1,27 @@
 import {observer} from "mobx-react-lite";
-import {DefaultDelete} from "../DefaultDelete";
-import {NavbarComponent} from "../Navbar";
+import {DefaultDelete} from "../../components/DefaultDelete";
+import {NavbarComponent} from "../../components/Navbar";
 import {useNavigate, useParams} from "react-router-dom";
-import {useToken} from "../Auth/AuthContext";
+import {useToken} from "../../utils/AuthContext";
 import {Container} from "react-bootstrap";
 
-export const TransactionDelete = observer(function TransactionDelete() {
-  const {id} = useParams();
-  const backLink: string = `/transactions/${id}`;
+export const AccountDelete = observer(function AccountDelete() {
   const auth = useToken();
   const navigate = useNavigate();
+  const {id} = useParams();
+  const backLink: string = `/accounts/${id}`;
   if (auth.getToken() === "") {
     navigate("/login");
   }
+
   return (
     <Container>
       <NavbarComponent />
       <DefaultDelete
         backLink={backLink}
-        returnPoint={`/transactions`}
         id={id ? id : ""}
-        deleteRequestUrl={`transactions`}
+        returnPoint={`/accounts`}
+        deleteRequestUrl={"accounts"}
       />
     </Container>
   );
