+++ expense_tracker_frontend/src/pages/Accounts/AccountEdit.tsx
@@ -1,20 +1,20 @@
 import {observer} from "mobx-react-lite";
 import {FormEvent, useEffect, useState} from "react";
-import {useToken} from "../Auth/AuthContext";
-import {NavbarComponent} from "../Navbar";
+import {useToken} from "../../utils/AuthContext";
+import {NavbarComponent} from "../../components/Navbar";
 import {useNavigate, useParams} from "react-router-dom";
 import {AuthAxios} from "../../utils/Network";
 import {Col, Form, Row, Container} from "react-bootstrap";
-import {SubmitButton} from "../SubmitButton";
+import {SubmitButton} from "../../components/SubmitButton";
 
-interface Tag {
+interface Account {
   id: number;
   name: string;
   desc: string;
   user: string;
 }
 
-export const TagEdit = observer(function TagEdit() {
+export const AccountEdit = observer(function AccountEdit() {
   const auth = useToken();
   const navigate = useNavigate();
   const {id} = useParams();
@@ -24,30 +24,28 @@ export const TagEdit = observer(function TagEdit() {
   const [name, setName] = useState("");
   const [desc, setDesc] = useState("");
   useEffect(() => {
-    AuthAxios.get(`tags?id=${id}`, auth.getToken())
+    AuthAxios.get(`accounts?id=${id}`, auth.getToken())
       .then((res) => {
-        const data: Tag = res.data[0];
+        const data: Account = res.data[0];
         setName(data.name);
         setDesc(data.desc);
       })
       .catch((err) => console.error(err));
   }, []);
-
   let bodyParameters = {
     id: id,
     Name: ``,
     Description: ``,
     action: "edit",
   };
-
   const submitHandler = (e: FormEvent<HTMLFormElement>) => {
     e.preventDefault();
     bodyParameters.Name = name;
     bodyParameters.Description = desc;
-    AuthAxios.post("tags", auth.getToken(), bodyParameters).catch((err) =>
+    AuthAxios.post("accounts", auth.getToken(), bodyParameters).catch((err) =>
       console.error(err),
     );
-    navigate(`/tags/${id}`);
+    navigate(`/accounts/${id}`);
   };
   return (
     <Container>
