+++ expense_tracker_frontend/src/pages/Auth/UserEdit.tsx
@@ -1,6 +1,6 @@
 import {observer} from "mobx-react-lite";
-import {useToken} from "./AuthContext";
-import {NavbarComponent} from "../Navbar";
+import {useToken} from "../../utils/AuthContext";
+import {NavbarComponent} from "../../components/Navbar";
 import {Container, Button} from "react-bootstrap";
 
 export const UserEdit = observer(function UserEdit() {
