+++ expense_tracker_frontend/src/pages/Presets/Presets.tsx
@@ -1,8 +1,8 @@
-import {NavbarComponent} from "../Navbar";
+import {NavbarComponent} from "../../components/Navbar";
 import React, {useEffect, useState} from "react";
 import {Link, useNavigate} from "react-router-dom";
-import {TableButton} from "../TableButton";
-import {useToken} from "../Auth/AuthContext";
+import {TableButton} from "../../components/TableButton";
+import {useToken} from "../../utils/AuthContext";
 import {observer} from "mobx-react-lite";
 import {AuthAxios} from "../../utils/Network";
 import {Col, Row, Container, Table, Button} from "react-bootstrap";
