+++ expense_tracker_frontend/src/pages/Tags/TagCreate.tsx
@@ -1,56 +1,40 @@
+import {NavbarEmpty} from "../../components/NavbarEmpty";
 import {observer} from "mobx-react-lite";
-import {FormEvent, useEffect, useState} from "react";
-import {useToken} from "../Auth/AuthContext";
-import {NavbarComponent} from "../Navbar";
-import {useNavigate, useParams} from "react-router-dom";
+import {FormEvent, useState} from "react";
+import {useToken} from "../../utils/AuthContext";
+import {NavbarComponent} from "../../components/Navbar";
+import {useNavigate} from "react-router-dom";
 import {AuthAxios} from "../../utils/Network";
 import {Col, Form, Row, Container} from "react-bootstrap";
-import {SubmitButton} from "../SubmitButton";
+import {SubmitButton} from "../../components/SubmitButton";
 
-interface Account {
-  id: number;
-  name: string;
-  desc: string;
-  user: string;
-}
-
-export const AccountEdit = observer(function AccountEdit() {
+export const TagCreate = observer(function TagCreate() {
   const auth = useToken();
   const navigate = useNavigate();
-  const {id} = useParams();
+  const [name, setName] = useState("");
+  const [desc, setDesc] = useState("");
   if (auth.getToken() === "") {
     navigate("/login");
   }
-  const [name, setName] = useState("");
-  const [desc, setDesc] = useState("");
-  useEffect(() => {
-    AuthAxios.get(`accounts?id=${id}`, auth.getToken())
-      .then((res) => {
-        const data: Account = res.data[0];
-        setName(data.name);
-        setDesc(data.desc);
-      })
-      .catch((err) => console.error(err));
-  }, []);
   let bodyParameters = {
-    id: id,
     Name: ``,
     Description: ``,
-    action: "edit",
+    action: "create",
   };
-  const submitHandler = (e: FormEvent<HTMLFormElement>) => {
+
+  const submitHandler = async (e: FormEvent<HTMLFormElement>) => {
     e.preventDefault();
     bodyParameters.Name = name;
     bodyParameters.Description = desc;
-    AuthAxios.post("accounts", auth.getToken(), bodyParameters).catch((err) =>
+    await AuthAxios.post("tags", auth.getToken(), bodyParameters).catch((err) =>
       console.error(err),
     );
-    navigate(`/accounts/${id}`);
+    navigate("/tags");
   };
   return (
     <Container>
       <NavbarComponent />
-      <h1>Edit</h1>
+      <h1>Create new tag</h1>
       <Form id="tag-create-form" onSubmit={submitHandler}>
         <Form.Group>
           <Row className="mb-3">
@@ -59,7 +43,6 @@ export const AccountEdit = observer(function AccountEdit() {
             </Col>
             <Col xs={8} sm={10}>
               <Form.Control
-                value={name}
                 type="text"
                 name="name"
                 key="id_name"
@@ -74,7 +57,6 @@ export const AccountEdit = observer(function AccountEdit() {
             </Col>
             <Col xs={8} sm={10}>
               <Form.Control
-                value={desc}
                 type="text"
                 name="description"
                 key="id_desc"
