+++ expense_tracker_frontend/src/pages/Transactions/Transaction.tsx
@@ -1,11 +1,15 @@
 import {observer} from "mobx-react-lite";
-import {NavbarComponent} from "../Navbar";
-import {TableButton} from "../TableButton";
-import {StaticField} from "../StaticField";
+import {NavbarComponent} from "../../components/Navbar";
+import {TableButton} from "../../components/TableButton";
+import {StaticField} from "../../components/StaticField";
 import React, {useEffect, useState} from "react";
-import {useToken} from "../Auth/AuthContext";
+import {useToken} from "../../utils/AuthContext";
 import {Link, useNavigate, useParams} from "react-router-dom";
-import {centsToString, formatDate, formatTimezone} from "../Tools";
+import {
+  centsToString,
+  formatDate,
+  formatTimezone,
+} from "../../components/Tools";
 import {AuthAxios} from "../../utils/Network";
 import {Col, Row, Table, Button, Alert, Container} from "react-bootstrap";
 import dayjs, {Dayjs} from "dayjs";
