+++ expense_tracker_frontend/src/pages/Transactions/TransactionCreate.tsx
@@ -1,6 +1,6 @@
 import {observer} from "mobx-react-lite";
-import {NavbarComponent} from "../Navbar";
-import {useToken} from "../Auth/AuthContext";
+import {NavbarComponent} from "../../components/Navbar";
+import {useToken} from "../../utils/AuthContext";
 import {useNavigate} from "react-router-dom";
 import {
   ChangeEvent,
@@ -11,8 +11,11 @@ import {
   useRef,
   useState,
 } from "react";
-import {SubmitButton} from "../SubmitButton";
-import {formatDateIso8601, formatDateTimeForInput} from "../Tools";
+import {SubmitButton} from "../../components/SubmitButton";
+import {
+  formatDateIso8601,
+  formatDateTimeForInput,
+} from "../../components/Tools";
 import {AuthAxios} from "../../utils/Network";
 import {
   Card,
@@ -26,7 +29,7 @@ import {
   Collapse,
 } from "react-bootstrap";
 import dayjs, {Dayjs} from "dayjs";
-import {TimezoneSelect} from "../TimezoneSelect";
+import {TimezoneSelect} from "../../components/TimezoneSelect";
 
 interface Preset {
   id: number;
