+++ expense_tracker_frontend/src/pages/Transactions/TransactionDelete.tsx
@@ -1,27 +1,26 @@
 import {observer} from "mobx-react-lite";
-import {DefaultDelete} from "../DefaultDelete";
-import {NavbarComponent} from "../Navbar";
+import {DefaultDelete} from "../../components/DefaultDelete";
+import {NavbarComponent} from "../../components/Navbar";
 import {useNavigate, useParams} from "react-router-dom";
-import {useToken} from "../Auth/AuthContext";
+import {useToken} from "../../utils/AuthContext";
 import {Container} from "react-bootstrap";
 
-export const AccountDelete = observer(function AccountDelete() {
+export const TransactionDelete = observer(function TransactionDelete() {
+  const {id} = useParams();
+  const backLink: string = `/transactions/${id}`;
   const auth = useToken();
   const navigate = useNavigate();
-  const {id} = useParams();
-  const backLink: string = `/accounts/${id}`;
   if (auth.getToken() === "") {
     navigate("/login");
   }
-
   return (
     <Container>
       <NavbarComponent />
       <DefaultDelete
         backLink={backLink}
+        returnPoint={`/transactions`}
         id={id ? id : ""}
-        returnPoint={`/accounts`}
-        deleteRequestUrl={"accounts"}
+        deleteRequestUrl={`transactions`}
       />
     </Container>
   );
